refactor result section components
